update ver. 11.50 in Lang.ja-JP.xaml
update ver. 11.50 Added rajan and new event items.
In addition, all items with the same name, such as Tetoru dolls and collectibles, have been modified, and jewels that have been described as stones have been modified by throwing stones.

Corrected other item names with the same name (such as history)

In addition, the phenomenon that freezes when multiple items with the same name are selected and dropped in Japanese.

Since the above phenomenon occurred, the corrections were made to the extent possible other than the items of update ver. 11.50.