update ver. 11.50 in Lang.ja-JP.xaml　part2
update ver. 11.50 Added rajan and new event items.
In addition, all items with the same name, such as Tetoru dolls and collectibles, have been modified, and jewels that have been described as stones have been modified by throwing stones.

Corrected other item names with the same name (such as history)

And we could not send it for 12 days before due to an error, so we will send it again.